Fix a non-64-bit-clean printf argument.
git-svn-id: svn://svn.tartarus.org/sgt@9848 cda61777-01e9-0310-a592-d414129be87e

+++ putty-spinoffs/unix/uxxtruss.c
@@ -7061,7 +7061,8 @@ int main(int argc, char **argv)
 				    set = &events_to_log;
 				else {
 				    fprintf(stderr, "xtruss: unknown keyword"
-					    " for '-e': '%.*s'\n", p-val, val);
+					    " for '-e': '%.*s'\n",
+                                            (int)(p-val), val);
 				    return 1;
 				}
 				p++;   /* skip '=' */
